% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{analysis_plot}
\alias{analysis_plot}
\title{Analysis plot}
\usage{
analysis_plot(cv, benchmark = T)
}
\arguments{
\item{cv}{list of object ppslp of the simulation}

\item{benchmark}{indicate whether to plot hastie}
}
\value{
mse
}
\description{
plot simulation results
}
\keyword{Principal}
\keyword{curve}
