% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.doublematrix}
\alias{as.doublematrix}
\title{function from GLODE}
\usage{
as.doublematrix(m)
}
\arguments{
\item{m}{matrix}
}
\description{
prepare some functions(from GLODE code)
}
\keyword{internal}
\keyword{norm}
