% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcgd.R
\name{bcgd}
\alias{bcgd}
\title{Block Coordinate descence}
\usage{
bcgd(
  m_ybar,
  m_bbar,
  m_x,
  m_ldtd,
  lambda_l,
  scaled_lambda,
  maxit = 1e+08,
  tol = 0.001
)
}
\arguments{
\item{m_ybar}{matrix Y}

\item{m_bbar}{matrix B}

\item{m_x}{matrix X}

\item{m_ldtd}{matrix D}

\item{lambda_l}{length parameter}

\item{scaled_lambda}{scaled length parameter}

\item{maxit}{maximum number of iterations, default to be 60}
}
\value{
\code{m_beta} matrix Beta
}
\description{
An group lasso solver
}
\examples{
library(ppslp)
m_y <- matrix(c(13, -4, 2, -4, 11, -2, 2, -2, 8, 10), 5, 2, byrow = TRUE)
s_k     <- 5L;
m_ybar  <- center_colmeans(m_y)
# initialize m_Theta
m_Theta <- f_initialization(m_y, s_k, w=6)
m_beta  <- diff(m_Theta) # Transform m_Theta to m_beta
# projection step
m_b     <- f_projection(m_y, m_Theta, m_beta)
m_bbar  <- center_colmeans(m_b);
lambda_s <- 1;
lambda_l <- 1;
bcgd(m_ybar, m_bbar, lambda_l, lambda_s)
}
\keyword{group}
\keyword{lasso}
