% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_ppslp.R
\name{cv_ppslp}
\alias{cv_ppslp}
\title{Cross Validation Tune Smoothing parameter
using grid search}
\usage{
cv_ppslp(
  m_y,
  s_k,
  l_range = 8,
  s_range = 4,
  folds = 5,
  lambda_l = 1,
  maxit = 100,
  init = NULL,
  thresh = 1e-04
)
}
\arguments{
\item{m_y}{your data matrix Y}

\item{s_k}{number of knots}

\item{l_range}{scaled length range(from 0 to 1):
either a vector of length parameters
or an integer indicating the range of the selection range
in exponential:
           exp(-range:-1)}

\item{s_range}{smoothing range(from 0 to inf):
either a vector of smoothing parameters
or an integer indicating the range of the selection range
in exponential:
           exp(-range:range)}

\item{folds}{the number of fold to split Y, default to be 5}

\item{maxit}{maximum number of iterations, default to be 60}

\item{init}{manual start point for initialization, default to be \code{NULL},
so initializae with first principal component curve}

\item{thresh}{tolerance or stopping criteria,
if change in the criteria less than this value,
the algorithm will terminate}
}
\value{
a fitted class ppslp using the selected parameters
}
\description{
Cross Validation Tune Smoothing parameter
using grid search
}
\keyword{Principal}
\keyword{curve}
