% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{cv_simulation}
\alias{cv_simulation}
\title{Grid Cross Validation simulation}
\usage{
cv_simulation(ncores = 16, s_n = 100L, s_k = 25L, s_q = 2L, ...)
}
\arguments{
\item{ncores}{number of cores used in simulation}

\item{s_n}{number of observations}

\item{s_k}{number of knots}

\item{s_q}{dimension}

\item{...}{other args for cv_grid(maxit, thresh, l_range, s_range, folds)}
}
\value{
\code{s_lambda_p} smoothing parameter
}
\description{
parallel simulation
}
\keyword{Principal}
\keyword{curve}
