% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_2nrm}
\alias{f_2nrm}
\title{The 2norm or Frobenius norm function}
\usage{
f_2nrm(x)
}
\arguments{
\item{x}{vector or matrix to calculate its 2norm or Frobenius norm}
}
\value{
2norm or Frobenius norm
}
\description{
The 2norm or Frobenius norm function
}
\keyword{internal}
\keyword{norm}
