% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{f_Animeppslp}
\alias{f_Animeppslp}
\title{Principal Curve with Length Penalty with animation}
\usage{
f_Animeppslp(
  m_Y,
  s_K,
  s_q,
  pen = 1,
  increment = 0.1,
  maxit = 60,
  s_lambdap = 2,
  init_w = 6,
  init_Theta = NULL,
  tol = 0.001
)
}
\arguments{
\item{m_Y}{your data matrix Y}

\item{s_K}{number of knots}

\item{s_q}{dimension of the data, default to be \code{ncol(m_Y)}}

\item{pen}{penalty term}

\item{increment}{increment step of length penalty parameter, default to be 0.1}

\item{maxit}{maximum number of iterations, default to be 60}

\item{s_lambdap}{smoother parameter, default to be 1}

\item{init_w}{initialization weights of the first principal component, default to be 6}

\item{init_Theta}{manual start point for initialization, default to be \code{NULL}, so initializae with first principal component curve}

\item{tol}{tolerance or stopping criteria, if change in coefficient is less than this value the algorithm will terminate}
}
\value{
\code{m_Theta} output from the last equation, plot the principal curve
}
\description{
Wrapper function allows you to visualize the fitting procedure of a principal curve with length penalty.
}
\keyword{Principal}
\keyword{curve}
