% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{f_lenpen}
\alias{f_lenpen}
\title{Evaluation}
\usage{
f_lenpen(m_Theta, lambda_l)
}
\arguments{
\item{m_Theta}{matrix m_Theta}

\item{s_lambda_L}{length parameter}
}
\value{
Length penalty
}
\description{
Length Penalty
}
\keyword{Evaluation}
\keyword{internal}
