% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_transformation.R
\name{f_mx}
\alias{f_mx}
\title{Conversion matrix}
\usage{
f_mx(s_k)
}
\arguments{
\item{s_k}{number of knots}
}
\value{
\code{m_x}
}
\description{
formulate m_x,
}
\keyword{Variable}
\keyword{internal}
\keyword{transformation}
