% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppslp.R
\name{f_ppslp}
\alias{f_ppslp}
\alias{lines.ppslp}
\alias{plot.ppslp}
\alias{points.ppslp}
\alias{whiskers}
\title{Principal Curve with Length Penalty}
\usage{
f_ppslp(
  m_y,
  s_k,
  lambda_s = 20,
  scaled_lam = 0.01,
  maxit = 1000,
  init = NULL,
  thresh = 0.001,
  plot_update = F,
  lambda_l = NULL
)

\method{lines}{ppslp}(pp, ...)

\method{plot}{ppslp}(pp, ...)

\method{points}{ppslp}(pp, ...)

whiskers(x, s, ...)
}
\arguments{
\item{m_y}{your data matrix y}

\item{s_k}{number of knots}

\item{lambda_s}{smoother parameter, default to be 1}

\item{scaled_lam}{scaled length parameter,
$=\lambda_l/\lambda_{\max}$, default to be 0.01}

\item{maxit}{maximum number of iterations, default to be 60}

\item{init}{either a previously fit principal curve,
 or starting value of matrix theta.
 If missing or NULL, then the first principal component is used.
If  then a circle is used as the start.}

\item{thresh}{convergence treshhold}

\item{plot_update}{plot the update}

\item{lambda_l}{length parameter, default to be NULL}

\item{s}{a parametrized curve, represented by a polygon.}
}
\value{
\code{out} An prinicipal curve opject includes
       \item{y}{matrix y}
       \item{fitted_y}{fitted y on principal curve}
       \item{B}{matrix B}
       \item{theta}{spline matrix theta}
       \item{iter}{number of iteration}
       \item{lambda_l}{length parameter}
       \item{lambda_s}{smoother parameter}
       \item{dist}{difference of theta}
       \item{converged}{A logical indicating whether the algorithm converged
             or not.}
       \item{num_iterations}{Number of iterations
                             completed before returning.}
       \item{call}{the call that created this object; allows it to be
             \code{updated()}.}
}
\description{
This function allows you to fit a principal curve with length penalty.
}
\examples{
\dontrun{
devtools::load_all() # our package
data("threeExample")
y <- cbind(threeExample$x, threeExample$y)
fit <- f_ppslp(y, k = 50, lambda_s = 10, lambda_l = 0.1)
plot(fit)
lines(fit)
points(fit)
whiskers(y, fit$fitted_y)
points(y, pch = 16, cex = 0.8, col = "red")
}
}
\keyword{Principal}
\keyword{curve}
