% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{optimization}
\alias{optimization}
\title{Optimization}
\usage{
optimization(m_ybar, m_b, m_x, m_ldtd, lambda_l, scaled_lambda)
}
\arguments{
\item{m_ybar}{matrix Y}

\item{m_b}{matrix B}

\item{m_x}{matrix X}

\item{m_ldtd}{matrix lam_p * D^TD}

\item{lambda_l}{length parameter}

\item{scaled_lambda}{scaled length parameter}
}
\value{
list \code{m_Theta}, status indicator for if it collaspe
}
\description{
Optimization
}
