% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{orig_criteria}
\alias{orig_criteria}
\title{Evaluation}
\usage{
orig_criteria(pp, new_data = NULL)
}
\arguments{
\item{pp}{an prinicipal curve opject includes
\item{Y}: matrix Y
\item{Theta}: spline matrix m_Theta
\item{iter} number of iteration
\item{lambda_l} length parameter
\item{lambda_s} smoother parameter
\item{dist} difference of m_Theta}

\item{new_data}{External data matrix}
}
\value{
\code{diff}, difference b/w coefficient m_Theta
}
\description{
Evaluate Distance
}
\keyword{Evaluation}
\keyword{internal}
