% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pp_plot}
\alias{pp_plot}
\title{Plot Principal Curve with Length Penalty}
\usage{
pp_plot(pp)
}
\arguments{
\item{pp}{an prinicipal curve opject includes
\item{Y}: matrix Y
\item{Theta}: spline matrix m_Theta
\item{iter} number of iteration
\item{lambda_l} length parameter
\item{lambda_s} smoother parameter
\item{dist} difference of m_Theta}
}
\description{
This function allows you to fit a principal curve with length penalty.
}
\keyword{Principal}
\keyword{curve}
