% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{spectExample}
\alias{spectExample}
\title{Example CT/SPECT from colon dataset}
\format{
An object of class \code{list} of length 8.
}
\usage{
data(spectExample)
}
\description{
A example data from the SPECT dataset which contains both the SPECT and CT image
and its start point and end point
}
\keyword{datasets}
