% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{theta_diff}
\alias{theta_diff}
\title{Evaluation}
\usage{
theta_diff(new_theta, old_theta)
}
\arguments{
\item{m_Y}{data matrix m_Y}

\item{m_Theta}{matrix m_Theta}
}
\value{
\code{diff}, difference b/w coefficient m_Theta
}
\description{
Evaluate the difference b/w coefficient m_Theta
}
\keyword{Evaluation}
