% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleData.R
\docType{data}
\name{threeExample}
\alias{threeExample}
\title{Example number from MNIST dataset}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(threeExample)
}
\description{
Number 3 data from MNIST dataset which contains the scaled coordinate of numbre 3
and its start point and end point
}
\keyword{datasets}
