% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{xy_fix}
\alias{xy_fix}
\title{Start and end point of the simulation Examples}
\usage{
xy_fix(case)
}
\arguments{
\item{case}{number indicates the examples(1-7):
1: 3/4 circle
2: 3/4 circle with noise
3: 3/4 circle with noise, initialize with the truth
4: 1/2 circle
5: 1/2 circle with noise
6: 1/2 circle with noise, initialize with the truth
7: the shape of number 3 from mnist dataset}
}
\value{
the start_end point for each case
}
\description{
Start and end point of the simulation Examples
}
\keyword{Principal}
\keyword{curve}
