% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{xy_truth}
\alias{xy_truth}
\title{The truth curve of the simulation Examples}
\usage{
xy_truth(case, n = 10000, q = 2)
}
\arguments{
\item{case}{number indicates the examples(1-7):
1: 3/4 circle
2: 3/4 circle with noise
3: 3/4 circle with noise, initialize with the truth
4: 1/2 circle
5: 1/2 circle with noise
6: 1/2 circle with noise, initialize with the truth
7: the shape of number 3 from mnist dataset}

\item{n}{number of the truth}

\item{q}{dimension of the simulation}
}
\value{
the start_end point for each case
}
\description{
The truth curve of the simulation Examples
}
\keyword{Principal}
\keyword{curve}
