% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{y_generator}
\alias{y_generator}
\title{Simulation Examples}
\usage{
y_generator(case, n = 100L, q = 2L, seeds = 999983)
}
\arguments{
\item{case}{number indicates the examples(1-7):
1: 3/4 circle
2: 3/4 circle with noise
3: 3/4 circle with noise, initialize with the truth
4: 1/2 circle
5: 1/2 circle with noise
6: 1/2 circle with noise, initialize with the truth
7: the shape of number 3 from mnist dataset}

\item{n}{number of obervation in the simulation}

\item{q}{dimension of the simulation}
}
\value{
the simulation examples
}
\description{
Simulation Examples
}
\keyword{Principal}
\keyword{curve}
